pyproject.toml
[tool.poetry]
name = "airbnb-clone-backend"
version = "0.1.0"
description = ""
authors = ["니꼬 <nico@nomadcoders.co>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1"
Pillow = "^9.2.0"
djangorestframework = "^3.14.0"
black = "^22.10.0"
pyjwt = "^2.6.0"
django-environ = "^0.9.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.6.0", allow-prereleases = true}

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
